move Submitter interface
